[CVAT][Exchange Oracle] Major refactoring (#533)
* Changed `network_id` type from str to int.
Escrow validators moved to `chain` folder.

* `network_id` replaced with `chain_id`

* changed folder structure
