--- packages/examples/cvat/exchange-oracle/src/modules/cvat/api_schema.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel
-from typing import Optional
-
-
-class CvatWebhook(BaseModel):
-    event: str
-    job: Optional[dict]
-    task: Optional[dict]
-    before_update: Optional[dict]
