+++ packages/examples/cvat/exchange-oracle/src/validators/signature.py
@@ -1,0 +1,3 @@
+# TODO: Validate webhook signature
+def validate_signature(signature: str):
+    pass
